requirements.txt
# Dependencies
# Install with: pip install -r requirements.txt

#IMPORTANT!! -- To run the app within your browser, you must enter "streamlit run app.py" inside your terminal
streamlit>=1.30
openai>=1.0.0
sentence-transformers>=2.6.0
faiss-cpu>=1.7.4
numpy>=1.24
pandas>=2.0
matplotlib>=3.7
seaborn>=0.13
scikit-learn>=1.3
networkx>=3.2
PyPDF2>=3.0.1
python-docx>=1.1.0
python-pptx>=0.6.22
datasets>=2.18.0
rouge-score>=0.1.2
transformers>=4.40.0
torch>=2.0.0
tqdm>=4.66